Renamed config to defaults
