+++ apps-rendering/src/components/MainMedia/MainMedia.defaults.ts
