+++ apps-rendering/src/components/MainMedia/index.tsx
@@ -7,7 +7,7 @@ import MainMediaVideo from 'components/MainMedia/MainMediaVideo';
 import { maybeRender } from 'lib';
 import { MainMedia, MainMediaKind } from 'mainMedia';
 import type { FC } from 'react';
-import { immersiveCaptionId } from './config';
+import { immersiveCaptionId } from './MainMedia.defaults';
 
 // ----- Component ----- //
